New handlers and methods implementations (#47)
* Wstępna wiadomość

* Test for brach conflict

* Test for resolving conflict 2

* Some new logic and some changes (will work on it)

* Jakaś logika handlerów (#44)

Adding handler logics

* Add basic testing infrastructure (#42)

* Change object schema and add startup class

* Add spectre.console lib and finish startup class

* finish

* Minor changes

* backup

* Work with some random test to prove that my sanity is not working, finish

* one more

* cleanup

* Prosze bez problemów

* Test commit

* updating pull request *don't merge* will still work on this branch

---------

+++ ColonySimulator.Backend/Migrations/20240515214822_init.cs
@@ -21,7 +21,7 @@ namespace ColonySimulator.Backend.Migrations
                     Strength = table.Column<double>(type: "REAL", nullable: false),
                     Agility = table.Column<double>(type: "REAL", nullable: false),
                     Gender = table.Column<int>(type: "INTEGER", nullable: false),
-                    ResourceConsumption = table.Column<double>(type: "REAL", nullable: false),
+                    ResourceConsumption = table.Column<int>(type: "INTEGER", nullable: false),
                     RequiredStrength = table.Column<double>(type: "REAL", nullable: false),
                     RequiredAgility = table.Column<double>(type: "REAL", nullable: false)
                 },
@@ -41,7 +41,7 @@ namespace ColonySimulator.Backend.Migrations
                     Strength = table.Column<double>(type: "REAL", nullable: false),
                     Agility = table.Column<double>(type: "REAL", nullable: false),
                     Gender = table.Column<int>(type: "INTEGER", nullable: false),
-                    ResourceConsumption = table.Column<double>(type: "REAL", nullable: false),
+                    ResourceConsumption = table.Column<int>(type: "INTEGER", nullable: false),
                     RequiredStrength = table.Column<double>(type: "REAL", nullable: false),
                     RequiredAgility = table.Column<double>(type: "REAL", nullable: false)
                 },
@@ -75,7 +75,7 @@ namespace ColonySimulator.Backend.Migrations
                     Strength = table.Column<double>(type: "REAL", nullable: false),
                     Agility = table.Column<double>(type: "REAL", nullable: false),
                     Gender = table.Column<int>(type: "INTEGER", nullable: false),
-                    ResourceConsumption = table.Column<double>(type: "REAL", nullable: false),
+                    ResourceConsumption = table.Column<int>(type: "INTEGER", nullable: false),
                     RequiredStrength = table.Column<double>(type: "REAL", nullable: false),
                     RequiredAgility = table.Column<double>(type: "REAL", nullable: false)
                 },
@@ -125,7 +125,7 @@ namespace ColonySimulator.Backend.Migrations
                     Strength = table.Column<double>(type: "REAL", nullable: false),
                     Agility = table.Column<double>(type: "REAL", nullable: false),
                     Gender = table.Column<int>(type: "INTEGER", nullable: false),
-                    ResourceConsumption = table.Column<double>(type: "REAL", nullable: false),
+                    ResourceConsumption = table.Column<int>(type: "INTEGER", nullable: false),
                     RequiredStrength = table.Column<double>(type: "REAL", nullable: false),
                     RequiredAgility = table.Column<double>(type: "REAL", nullable: false)
                 },
@@ -177,7 +177,7 @@ namespace ColonySimulator.Backend.Migrations
                     Strength = table.Column<double>(type: "REAL", nullable: false),
                     Agility = table.Column<double>(type: "REAL", nullable: false),
                     Gender = table.Column<int>(type: "INTEGER", nullable: false),
-                    ResourceConsumption = table.Column<double>(type: "REAL", nullable: false),
+                    ResourceConsumption = table.Column<int>(type: "INTEGER", nullable: false),
                     RequiredStrength = table.Column<double>(type: "REAL", nullable: false),
                     RequiredAgility = table.Column<double>(type: "REAL", nullable: false)
                 },
@@ -197,7 +197,7 @@ namespace ColonySimulator.Backend.Migrations
                     Strength = table.Column<double>(type: "REAL", nullable: false),
                     Agility = table.Column<double>(type: "REAL", nullable: false),
                     Gender = table.Column<int>(type: "INTEGER", nullable: false),
-                    ResourceConsumption = table.Column<double>(type: "REAL", nullable: false),
+                    ResourceConsumption = table.Column<int>(type: "INTEGER", nullable: false),
                     RequiredStrength = table.Column<double>(type: "REAL", nullable: false),
                     RequiredAgility = table.Column<double>(type: "REAL", nullable: false)
                 },
